PSReadLineProfile.ps1: add some checks for older versions to avoid some errors

+++ PSReadLineProfile.ps1
@@ -27,22 +27,29 @@ Set-PSReadLineKeyHandler -Chord 'DownArrow' -Function 'HistorySearchForward'
 #
 # try to make different OSes behave the same:
 #
-$kh = Get-PSReadLineKeyHandler -Chord 'Ctrl+h'
-if ($kh -and $kh.Function -eq 'BackwardDeleteChar') {
-	# windows has Ctrl+h bound to Backspace, so remove that; probably don't have to do this since we're rebinding below, but...
-	Remove-PSReadLineKeyHandler -Chord 'Ctrl+h'
-}
-$kh = Get-PSReadLineKeyHandler -Chord 'Alt+F7'
-if ($kh -and $kh.Function -eq 'ClearHistory') {
-	# not really sure what this does, don't want to hit by accident:
-	Remove-PSReadLineKeyHandler -Chord 'Alt+F7'
+if ($mod.Version -ge ([System.Version]'2.0.1')) {	# -Chord added with v2.0.1; don't know how to search for it otherwise, and ... whatever
+	$kh = Get-PSReadLineKeyHandler -Chord 'Ctrl+h'
+	if ($kh -and $kh.Function -eq 'BackwardDeleteChar') {
+		# windows has Ctrl+h bound to Backspace, so remove that; probably don't have to do this since we're rebinding below, but...
+		Remove-PSReadLineKeyHandler -Chord 'Ctrl+h'
+	}
+	$kh = Get-PSReadLineKeyHandler -Chord 'Alt+F7'
+	if ($kh -and $kh.Function -eq 'ClearHistory') {
+		# not really sure what this does, don't want to hit by accident:
+		Remove-PSReadLineKeyHandler -Chord 'Alt+F7'
+	}
 }
 Set-PSReadLineKeyHandler -Chord 'Ctrl+h' -Function 'BackwardKillWord'				# in *nix shells, this is same as Ctrl+Backspace, so bind it to this too
 Set-PSReadLineKeyHandler -Chord 'Ctrl+u' -Function 'RevertLine'						# used in *nix shells, so add it; Escape also bound by default
 Set-PSReadLineKeyHandler -Chord 'Ctrl+a' -Function 'BeginningOfLine'				# used in *nix shells, so keep it for consistency; Home also bound by default
 Set-PSReadLineKeyHandler -Chord 'Ctrl+e' -Function 'EndOfLine'						# used in *nix shells, so keep it for consistency; End also bound by default
-Set-PSReadLineKeyHandler -Chord 'Ctrl+k' -Function 'ForwardDeleteInput'				# used in *nix shells, so keep it for consistency; Ctrl+End also bound
-Set-PSReadLineKeyHandler -Chord 'Ctrl+End' -Function 'ForwardDeleteInput'			# on by default for Win but not others
+if ($mod.Version -ge ([System.Version]'2.2.2')) {	# ForwardDeleteInput added with 2.2.2 (but think it's just a renamed ForwardDeleteLine ???, so that's we'll use for older)
+	Set-PSReadLineKeyHandler -Chord 'Ctrl+k' -Function 'ForwardDeleteInput'				# used in *nix shells, so keep it for consistency; Ctrl+End also bound
+	Set-PSReadLineKeyHandler -Chord 'Ctrl+End' -Function 'ForwardDeleteInput'			# on by default for Win but not others
+} else {
+	Set-PSReadLineKeyHandler -Chord 'Ctrl+k' -Function 'ForwardDeleteLine'				# used in *nix shells, so keep it for consistency; Ctrl+End also bound
+	Set-PSReadLineKeyHandler -Chord 'Ctrl+End' -Function 'ForwardDeleteLine'			# on by default for Win but not others
+}
 Set-PSReadLineKeyHandler -Chord 'Ctrl+Delete' -Function 'KillWord'					# on by default for Win but not others
 Set-PSReadLineKeyHandler -Chord 'Ctrl+Spacebar' -Function 'MenuComplete'			# on by default for Win but not others
 Set-PSReadLineKeyHandler -Chord 'PageDown' -Function 'ScrollDisplayDown'			# on by default for Win but not others
